envs/shop_sku_manager/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-shop-sku-manager"
version = "0.1.0"
description = "Shop SKU Manager Environment for OpenEnv - inventory management for retail agents"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[tool.setuptools]
include-package-data = true
packages = ["shop_sku_manager", "shop_sku_manager.server"]
package-dir = { "shop_sku_manager" = ".", "shop_sku_manager.server" = "server" }
